Clean up stale PVCs in workspace controller to avoid race condition
